+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [0.5.0]
 
 * Dynamic value in [DockingItem] class
-* Closable [DockingItem]
+* Non-Closable [DockingItem]
 
 ## [0.4.0]